[refactor] 예약 리펙토링 * 좌석에 낙관적 락 적용 * ReservationDetailResponse 에서 Info들 dto로 분리 * 낙관적 락 예외 처리 핸들러 추가
