+++ backend/src/main/java/com/pickgo/domain/performance/dto/PerformanceInfo.java
@@ -1,0 +1,21 @@
+package com.pickgo.domain.performance.dto;
+
+import com.pickgo.domain.performance.entity.Performance;
+
+public record PerformanceInfo(
+        String name,
+        String poster,
+        String runtime,
+        String type,
+        String state
+) {
+    public static PerformanceInfo from(Performance p) {
+        return new PerformanceInfo(
+                p.getName(),
+                p.getPoster(),
+                p.getRuntime(),
+                p.getType().getValue(),
+                p.getState().getValue()
+        );
+    }
+}
