+++ backend/src/main/java/com/pickgo/domain/performance/dto/VenueInfo.java
@@ -1,0 +1,15 @@
+package com.pickgo.domain.performance.dto;
+
+import com.pickgo.domain.performance.entity.Venue;
+
+public record VenueInfo(
+        String name,
+        String address
+) {
+    public static VenueInfo from(Venue venue) {
+        return new VenueInfo(
+                venue.getName(),
+                venue.getAddress()
+        );
+    }
+}
